requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
Django==3.2
django-bootstrap5==23.3
django-cors-headers==4.2.0
django-redis==5.3.0
django-rest-framework==0.1.0
django-simpleui==2023.8.28
djangorestframework==3.14.0
PyMySQL==1.1.0
pytz==2023.3
redis==5.0.0
sqlparse==0.4.4
typing_extensions==4.7.1
